Perf refactoring 2 some perfections added

+++ vkAlissia-2/vkAlissia-2/Screens/Friends/FriendsViewController+TableView.swift
@@ -12,7 +12,7 @@ import UIKit
 
 extension FriendsViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        CGFloat(FriendCell.publicCellHeight)
+        FriendCell.publicCellHeight
     }
     
     func numberOfSections(in tableView: UITableView) -> Int {
